config: extract policy compilation to compiler_security_policy.go
Part B of #4121 (pure code motion). Moves the policy-compilation
cluster verbatim out of compiler_security.go:

  - compilePolicies
  - normalizePolicyAddrToken / normalizePolicyAddrTokens
  - policyMatchChildren / policyThenChildren / policyThenActionNodes
    (the #3842 duplicate-match/then-block accumulation helpers)
  - compilePolicy (the #4121 part-A firewallMatchValues SSOT reads
    shipped in #4133 travel with the function unchanged)
  - recognizedCollapsedDenyToken / applyCollapsedDenyModifiers
    (the #3141 flat-collapsed `then deny <modifier>` wiring)

Function bodies are byte-identical to the origin/master original. All
functions stay in package config; the strict match/then gates in the
sibling validator files (validatePolicyMatchLeavesStrict,
validatePolicyThenDenyStrict, ...) keep resolving these helpers across
the same package with no import change.

Validation: gofmt -l clean, go build ./pkg/config/ green.
